@@ -5,46 +5,63 @@
   "scripts": {
     "serve": "vue-cli-service serve",
     "build": "vue-cli-service build",
+    "test:unit": "vue-cli-service test:unit",
     "lint": "vue-cli-service lint",
-    "server:build": "tsc",
-    "server:watch": "tsc -w",
-    "test:unit": "vue-cli-service test:unit"
+    "server:build": "tsc -p server/tsconfig-server.json",
+    "server:run": "node build-server/index.js",
+    "server:watch": "tsc -p server/tsconfig-server.json -w"
   },
   "dependencies": {
+    "@types/bcrypt": "^5.0.0",
     "@types/express": "^4.17.2",
+    "@types/fs-extra": "^7.0.0",
+    "@types/js-yaml": "^4.0.4",
+    "@types/jsonwebtoken": "^8.5.5",
+    "@types/lodash": "^4.14.176",
+    "@types/node-fetch": "^2.5.12",
     "@types/sockjs": "^0.3.31",
+    "@types/sockjs-client": "^1.5.1",
     "await-lock": "^2.0.1",
+    "bcrypt": "^5.0.1",
     "body-parser": "^1.19.0",
-    "core-js": "^2.6.5",
+    "core-js": "^3.6.5",
     "epic-spinners": "^1.1.0",
     "express": "^4.17.1",
     "fs-extra": "^8.1.0",
-    "js-yaml": "^3.13.1",
+    "js-yaml": "^4.1.0",
+    "jsonwebtoken": "^8.5.1",
     "lodash": "^4.17.19",
     "moment": "^2.24.0",
-    "node-fetch": "^2.6.1",
+    "node-fetch": "^2.6.6",
     "reconnecting-websocket": "^4.1.10",
     "sockjs": "^0.3.19",
     "sockjs-client": "^1.3.0",
     "terser": "^4.0.0",
-    "vue": "^2.6.10",
-    "vue-slider-component": "^3.0.33",
-    "vuex": "^3.0.1",
-    "vuex-persistedstate": "^2.5.4"
+    "vue": "^3.2.21",
+    "vue-slider-component": "^4.0.0-beta.4",
+    "vue-types": "^4.1.1",
+    "vue-when": "^2.0.0",
+    "vuex": "^4.0.2",
+    "vuex-persistedstate": "^4.1.0"
   },
   "devDependencies": {
-    "@vue/cli-plugin-babel": "^3.8.0",
-    "@vue/cli-plugin-eslint": "^3.8.0",
-    "@vue/cli-plugin-unit-mocha": "^3.8.0",
-    "@vue/cli-service": "^3.8.0",
-    "@vue/test-utils": "1.0.0-beta.29",
-    "babel-eslint": "^10.0.1",
-    "chai": "^4.1.2",
-    "eslint": "^5.16.0",
-    "eslint-plugin-vue": "^5.0.0",
-    "sass": "^1.18.0",
-    "sass-loader": "^7.1.0",
-    "typescript": "^3.7.2",
-    "vue-template-compiler": "^2.6.10"
+    "@types/chai": "^4.2.11",
+    "@types/mocha": "^5.2.4",
+    "@typescript-eslint/eslint-plugin": "^4.18.0",
+    "@typescript-eslint/parser": "^4.18.0",
+    "@vue/cli-plugin-babel": "~4.5.15",
+    "@vue/cli-plugin-eslint": "~4.5.15",
+    "@vue/cli-plugin-typescript": "~4.5.0",
+    "@vue/cli-plugin-unit-mocha": "~4.5.15",
+    "@vue/cli-service": "~4.5.15",
+    "@vue/eslint-config-typescript": "^7.0.0",
+    "babel-eslint": "^10.1.0",
+    "chokidar": "^3.5.2",
+    "eslint": "^6.7.2",
+    "eslint-plugin-vue": "^6.2.2",
+    "fsevents": "^2.3.2",
+    "sass": "^1.43.4",
+    "sass-loader": "^7.3.1",
+    "typescript": "^4.4.4"
   }
 }
