@@ -1,5 +1,5 @@
 import { BaseCameraService } from './BaseCameraService';
-import * as fetch from 'node-fetch';
+import fetch from 'node-fetch';
 
 export interface IpeyeBackendConfig {
   streamType: string;
@@ -15,7 +15,7 @@ export class IpeyeBackend extends BaseCameraService {
 
   async fetchStreamUrl(cameraId: string) {
     const response = await fetch('http://api.ipeye.ru/device/url/' + this.streamType + '/' + cameraId);
-    const { message, status } = await response.json();
+    const { message, status } = await response.json() as any;
     if (!status) {
       throw new Error('IpEye API error: ' + message);
     }
