@@ -1,13 +1,14 @@
 import { Camera } from '../Camera';
 import { CameraStateStorage } from '../storage/CameraStateStorage';
 import { WebsocketServer } from './WebsocketServer';
+import express from 'express';
 
 export function updateCamera(
   cameras: ReadonlyArray<Camera>,
   cameraStateStorage: CameraStateStorage,
   websocketServer: WebsocketServer,
 ) {
-  return async (req, res) => {
+  return async (req: express.Request, res: express.Response) => {
     if ('poweredOff' in req.body) {
       for (const camera of cameras) {
         if (camera.id === req.params.id) {
