@@ -1,14 +1,16 @@
 import * as yaml from 'js-yaml';
 import * as path from 'path';
 import * as fs from 'fs-extra';
-import * as merge from 'lodash/merge';
+import merge from 'lodash/merge';
 
 export function readConfig() {
-  const defaults = yaml.safeLoad(fs.readFileSync(
-    path.resolve(__dirname, '..', 'config.defaults.yaml')
+  const defaults = yaml.load(fs.readFileSync(
+    path.resolve(__dirname, '..', 'config.defaults.yaml'),
+    { encoding: 'utf8' },
   ));
-  const userConfig = yaml.safeLoad(fs.readFileSync(
-    path.resolve(__dirname, '..', 'config.yaml')
+  const userConfig = yaml.load(fs.readFileSync(
+    path.resolve(__dirname, '..', 'config.yaml'),
+    { encoding: 'utf8' },
   ));
-  return merge(defaults, userConfig);
+  return merge(defaults, userConfig) as any;
 }
