@@ -1,342 +1,269 @@
 <template>
-  <div class='camera-grid' :style='{ fontSize }'>
-    <div v-for='y in grid.rows' class='row' :key='y'>
-      <div v-for='x in grid.columns' :class='["column", { "powered-off": isPoweredOff(x, y) }]' :key='x'>
-        <img v-if='cameraExists(x, y)'
-             :src='source(x, y)' @load='onLoaded(x, y)' />
-        <div v-if='cameraExists(x, y)' @click='openFullscreen($event, x, y)'
-             :class='{ reloaded: isReloaded(x, y), info: true }'>
-          <div class='text'>
-            <span class='name'>{{ camera(x, y).name }}</span>
-            <span v-if='!isPoweredOff(x, y) && hasSingleError(x, y)' class='error doubt'>?</span>
-            <span v-if='!isPoweredOff(x, y) && hasErrors(x, y)' class='error'>✗</span>
-          </div>
-          <div class='time'>{{ lastUpdated(x, y) }}</div>
-          <div class='toggler'>
-            <input type='checkbox' :checked='!camera(x, y).isPoweredOff' @click.stop
-              @change="toggleCameraPoweredOff(camera(x, y))" />
-          </div>
-          <div v-if='camera(x, y).loading' class='loading'>
-            <hollow-dots-spinner
-              :animation-duration="1500"
-              :dot-size="dotSize"
-              :dots-num="3"
-              color="#ffffff"
-            />
-          </div>
-        </div>
-      </div>
+  <div class="camera-grid" :style="{ fontSize }">
+    <div v-for="(row, rowIndex) in cells" :key="rowIndex" class="camera-grid__row">
+      <template v-for="(camera, cameraIndex) in row">
+        <CameraGridCell
+          v-if="camera"
+          :key="camera.id"
+          class="camera-grid__camera"
+          :camera="camera"
+          :dot-size="dotSize"
+          @click="openFullscreen($event, camera)"
+        />
+        <div v-else :key="cameraIndex" class="camera-grid__placeholder" />
+      </template>
     </div>
-    <transition name='fullscreen' @after-enter='fullscreen.open = true'>
-      <fullscreen-video v-if='fullscreen' :camera='fullscreen.camera' :style='fullscreen.style'
-                        @close='closeFullscreen' :open='fullscreen && fullscreen.open'
-                        :imgSrc='fullscreen.imgSrc' />
-    </transition>
+    <Transition
+      name="fullscreen-transition"
+      @enter="onTransitionEnter"
+      @after-enter="afterTransitionEnter"
+      @leave="onTransitionLeave"
+    >
+      <FullscreenVideo
+        v-if="selectedCamera"
+        class="camera-grid__fullscreen-video"
+        :camera="selectedCamera"
+        :open="animationComplete"
+        :img-src="fullscreenPreview"
+        @close="closeFullscreen"
+      />
+    </Transition>
   </div>
 </template>
 
-<script>
-import FullscreenVideo from './FullscreenVideo';
-import { HollowDotsSpinner } from 'epic-spinners';
-import moment from 'moment';
-import { mapState, mapActions } from 'vuex';
+<script lang="ts">
+import FullscreenVideo from './FullscreenVideo.vue';
+import { computed, defineComponent, nextTick, onUnmounted, Ref, ref } from 'vue';
+import { useStore } from 'vuex';
+import CameraGridCell from '@/CameraGridCell.vue';
+import { useCameraPreview } from '@/helpers';
+import { Camera } from '@/types';
 
 const IMAGE_ASPECT_RATIO = 16/9;
 
-export default {
-  components: {
-    FullscreenVideo, HollowDotsSpinner
-  },
-  computed: {
-    ...mapState(['cameras', 'currentTime', 'settings']),
-    cellSize () {
-      const cellWidth = this.width / this.grid.columns;
-      const cellHeight = this.height / this.grid.rows;
-      return Math.sqrt(cellWidth * cellWidth + cellHeight * cellHeight);
-    },
-    dotSize () {
-      return this.cellSize / 15;
-    },
-    fontSize () {
-      return String(this.cellSize * this.settings.fontSize / 100) + 'px';
-    },
-    grid () {
-      const variants = [];
-      for (let rowCount = 1; rowCount < this.cameras.length; rowCount++) {
-        const columnCount = Math.ceil(this.cameras.length / rowCount);
-        const hasEmptyRows = rowCount * columnCount - this.cameras.length >= columnCount;
-        if (!hasEmptyRows) {
-          variants.push([columnCount, rowCount]);
-        }
-      }
-      const targetAR = this.width / this.height;
-      let best = null;
-      let bestValue = null;
-      for (const variant of variants) {
-        const ar = variant[0] / variant[1] * IMAGE_ASPECT_RATIO;
-        const dist = Math.abs(ar - targetAR);
+type Dimensions = [Ref<number>, Ref<number>];
 
-        if (!best || dist < bestValue) {
-          best = variant;
-          bestValue = dist;
-        }
+function useScreenDimensions() {
+  const width = ref(window.innerWidth);
+  const height = ref(window.innerHeight);
+
+  function updateDimensions() {
+    width.value = window.innerWidth;
+    height.value = window.innerHeight;
+  }
+
+  window.addEventListener('resize', updateDimensions);
+
+  onUnmounted(() => {
+    window.removeEventListener('resize', updateDimensions);
+  });
+
+  return [width, height] as Dimensions;
+}
+
+function useGrid(dimensions: Dimensions, count: Ref<number>, cellAspectRatio: Ref<number>) {
+  const [width, height] = dimensions;
+
+  return computed(() => {
+    const variants = [];
+    for (let rowCount = 1; rowCount < count.value; rowCount++) {
+      const columnCount = Math.ceil(count.value / rowCount);
+      const hasEmptyRows = rowCount * columnCount - count.value >= columnCount;
+      if (!hasEmptyRows) {
+        variants.push([columnCount, rowCount] as [number, number]);
       }
-      return {
-        columns: best[0],
-        rows: best[1]
-      };
     }
-  },
-  data: function () {
+    const targetAR = width.value / height.value;
+    let best = null;
+    let bestValue = null;
+    for (const variant of variants) {
+      const ar = variant[0] / variant[1] * cellAspectRatio.value;
+      const dist = Math.abs(ar - targetAR);
+
+      if (!best || (bestValue !== null && dist < bestValue)) {
+        best = variant;
+        bestValue = dist;
+      }
+    }
+    return best!;
+  });
+}
+
+function useCellSize(dimensions: Dimensions, grid: Ref<[number, number]>) {
+  const cellWidth = computed(() => dimensions[0].value / grid.value[0]);
+  const cellHeight = computed(() => dimensions[1].value / grid.value[1]);
+  return computed(() => Math.sqrt(cellWidth.value ** 2 + cellHeight.value ** 2));
+}
+
+function useDotSize(cellSize: Ref<number>) {
+  return computed(() => cellSize.value / 15);
+}
+
+function useFontSize(cellSize: Ref<number>, fontSize: Ref<number>) {
+  return computed(() => String(cellSize.value * fontSize.value / 100) + 'px');
+}
+
+function useFullscreen(dimensions: Dimensions) {
+  const animationComplete = ref(false);
+  const cellRect = ref(null as DOMRect | null)
+  const selectedCamera = ref(null as Camera | null);
+
+  async function closeFullscreen() {
+    animationComplete.value = false;
+    // wait for next tick so that iframe could be closed before animation starts
+    // TODO: check whether this is still relevant with vue 3
+    await nextTick();
+    selectedCamera.value = null;
+  }
+
+  function getStartingPositionStyle() {
+    const rect = cellRect.value!;
     return {
-      fullscreen: null,
-      height: window.innerHeight,
-      loadedVersions: {},
-      width: window.innerWidth
+      bottom: `${dimensions[1].value - rect.top - rect.height}px`,
+      left: `${rect.left}px`,
+      right: `${dimensions[0].value - rect.left - rect.width}px`,
+      top: `${rect.top}px`
+    };
+  }
+
+  function openFullscreen(event: MouseEvent, camera: Camera) {
+    const rect = (event.currentTarget as HTMLElement).getBoundingClientRect();
+    selectedCamera.value = camera;
+    cellRect.value = rect;
+  }
+
+  function afterTransitionEnter() {
+    animationComplete.value = true;
+  }
+
+  function onceTransitionStarts(element: HTMLElement, callback: () => void) {
+    const onTransitionStart = () => {
+      callback();
+      element.removeEventListener('transitionstart', onTransitionStart);
+    }
+
+    element.addEventListener('transitionstart', onTransitionStart);
+  }
+
+  function onTransitionLeave(element: HTMLElement) {
+    onceTransitionStarts(element, () => {
+      Object.assign(element.style, getStartingPositionStyle());
+    });
+  }
+
+  function onTransitionEnter(element: HTMLElement) {
+    Object.assign(element.style, getStartingPositionStyle());
+    onceTransitionStarts(element, () => {
+      Object.assign(element.style, {
+        bottom: null,
+        left: null,
+        right: null,
+        top: null,
+      });
+    });
+  }
+
+  const fullscreenPreview = useCameraPreview(selectedCamera);
+
+  return {
+    afterTransitionEnter,
+    animationComplete,
+    closeFullscreen,
+    fullscreenPreview,
+    onTransitionEnter,
+    onTransitionLeave,
+    openFullscreen,
+    selectedCamera,
+  };
+}
+
+export default defineComponent({
+  components: {
+    CameraGridCell,
+    FullscreenVideo,
+  },
+  setup() {
+    const store = useStore();
+
+    const cameras = computed(() => store.state.cameras);
+    const settings = computed(() => store.state.settings);
+
+    const dimensions = useScreenDimensions();
+    const grid = useGrid(dimensions, computed(() => cameras.value.length), ref(IMAGE_ASPECT_RATIO));
+    const cellSize = useCellSize(dimensions, grid);
+    const dotSize = useDotSize(cellSize);
+    const fontSize = useFontSize(cellSize, computed(() => settings.value.fontSize));
+
+    const fullscreen = useFullscreen(dimensions);
+
+    const cells = computed(() => {
+      const rows = [];
+      const maxIndex = cameras.value?.length;
+
+      for (let r = 0; r < grid.value[1]; r++) {
+        const row = [];
+        for (let c = 0; c < grid.value[0]; c++) {
+          const index = grid.value[0] * r + c;
+          if (index >= maxIndex) {
+            row.push(null);
+          } else {
+            row.push(cameras.value[index]);
+          }
+        }
+        rows.push(row);
+      }
+
+      return rows;
+    });
+
+    return {
+      cameras,
+      cells,
+      dotSize,
+      fontSize,
+      grid,
+      ...fullscreen,
     };
   },
-  destroyed () {
-    window.removeEventListener('resize', this.updateDimensions);
-  },
-  methods: {
-    ...mapActions(['toggleCameraPoweredOff']),
-    camera (x, y) {
-      return this.cameras[this.cameraIndex(x, y)];
-    },
-    cameraExists (x, y) {
-      return this.cameraIndex(x, y) < this.cameras.length;
-    },
-    cameraIndex (x, y) {
-      return this.grid.columns * (y - 1) + x - 1;
-    },
-    closeFullscreen () {
-      this.fullscreen.open = false;
-      this.$nextTick(() => {
-        this.fullscreen = null;
-      });
-    },
-    hasErrors (x, y) {
-      return this.camera(x, y).failureCounter > 1;
-    },
-    hasSingleError (x, y) {
-      return this.camera(x, y).failureCounter === 1;
-    },
-    isPoweredOff (x, y) {
-      const camera = this.camera(x, y);
-      return camera && camera.isPoweredOff;
-    },
-    isReloaded (x, y) {
-      const index = this.cameraIndex(x, y);
-      const camera = this.cameras[index];
-      if (!(index in this.loadedVersions)) {
-        this.$set(this.loadedVersions, index, null);
-      }
-      return this.loadedVersions[index] === camera.imageVersion && camera.imageVersion > 1;
-    },
-    lastUpdated (x, y) {
-      const camera = this.camera(x, y);
-      if (!camera.lastUpdated) {
-        return 'не обновлялась';
-      }
-      return moment.min(camera.lastUpdated, this.currentTime).from(this.currentTime);
-    },
-    onLoaded (x, y) {
-      const index = this.cameraIndex(x, y);
-      this.loadedVersions[index] = this.cameras[index].imageVersion;
-    },
-    openFullscreen (event, x, y) {
-      const camera = this.camera(x, y);
-      const rect = event.currentTarget.getBoundingClientRect();
-      const style = {
-        bottom: `${this.height - rect.top - rect.height}px`,
-        left: `${rect.left}px`,
-        right: `${this.width - rect.left - rect.width}px`,
-        top: `${rect.top}px`
-      };
-      this.fullscreen = { camera, imgSrc: this.source(x, y), open: false, style };
-    },
-    source (x, y) {
-      const { imageVersion, id } = this.cameras[this.cameraIndex(x, y)];
-      return `/snapshots/${id}/output.jpg?v=` + imageVersion;
-    },
-    updateDimensions () {
-      this.height = window.innerHeight;
-      this.width = window.innerWidth;
-    }
-  },
-  mounted () {
-    this.$nextTick(() => {
-      window.addEventListener('resize', this.updateDimensions);
-    })
-  }
-}
+});
 </script>
 
-<style scoped>
-  @keyframes flash {
-    0% {
-      background-color: rgba(255, 255, 255, 0);
-    }
+<style lang="scss" scoped>
+.camera-grid {
+  display: flex;
+  flex-direction: column;
+  height: 100%;
+  width: 100%;
 
-    1% {
-      background-color: rgba(255, 255, 255, 1);
-    }
-
-    5% {
-      background-color: rgba(255, 255, 255, 0.3);
-    }
-
-    10% {
-      background-color: rgba(255, 255, 255, 1);
-    }
-
-    15% {
-      background-color: rgba(255, 255, 255, 0.3);
-    }
-
-    20% {
-      background-color: rgba(255, 255, 255, 1);
-    }
-    
-    100% {
-      background-color: rgba(255, 255, 255, 0);
-    }
-  }
-
-  .camera-grid {
-    display: flex;
-    flex-direction: column;
-    height: 100%;
-    width: 100%;
-  }
-
-  .row {
+  &__row {
     display: flex;
     flex: 1;
     flex-direction: row;
   }
 
-  .column {
-    align-items: center;
-    display: flex;
+  &__camera, &__placeholder {
     flex: 1;
-    overflow: hidden;
-    position: relative;
   }
 
-  .column.powered-off .info {
-    background-image: url('./off.svg');
-    background-position: center center;
-    background-repeat: no-repeat;
-    background-size: 50% 70%;
-    opacity: 0.5;
-  }
-
-  img {
-    bottom: 0;
-    height: 100%;
-    left: 0;
-    position: absolute;
-    right: 0;
-    top: 0;
-    width: 100%;
-  }
-
-  .info.reloaded {
-    animation: flash 1s linear;
-  }
-
-  .text {
-    color: white;
-    text-shadow: #000 0 0 5px;
-  }
-
-  .error {
-    color: red;
-  }
-
-  .error.doubt {
-    color: yellow;
-    font-size: 1.2em;
-  }
-
-  .ok {
-    color: rgb(0, 219, 0);
-  }
-
-  .error, .ok {
-    pointer-events: none;
-    user-select: none;
-  }
-
-  .info {
+  &__fullscreen-video {
     bottom: 0;
     left: 0;
     position: absolute;
     right: 0;
     top: 0;
-  }
-
-  .info .loading {
-    align-items: center;
-    background: rgba(0, 153, 255, 0.658);
-    display: flex;
-    height: 100%;
-    justify-content: center;
-    width: 100%;
-  }
-
-  .info .text {
-    bottom: 0;
-    left: 5px;
-    position: absolute;
-  }
-
-  .info .time {
-    color: white;
-    font-size: 0.4em;
-    left: 5px;
-    opacity: 0;
-    position: absolute;
-    text-shadow: #000 0 0 2px;
-    top: 5px;
-    transition: opacity 0.2s linear;
-  }
-
-  .info:hover .time {
-    opacity: 1;
-  }
-
-  .toggler {
-    bottom: 5px;
-    height: auto;
-    position: absolute;
-    right: 5px;
-  }
-
-  .toggler input {
-    display: block;
-  }
-
-  .fullscreen-video {
-    position: absolute;
     transition: bottom 0.5s ease, left 0.5s ease, right 0.5s ease, top 0.5s ease, opacity 0.5s ease;
   }
+}
 
-  .fullscreen-enter-active, .fullscreen-leave {
+.fullscreen-transition {
+  &-enter-to, &-leave-from {
     opacity: 1;
   }
 
-  .fullscreen-enter, .fullscreen-leave-active {
+  &-enter-from, &-leave-to {
     opacity: 0;
   }
-
-  .fullscreen-video:not(.fullscreen-enter):not(.fullscreen-leave-active) {
-    bottom: 0 !important;
-    left: 0 !important;
-    right: 0 !important;
-    top: 0 !important;
-  }
+}
 </style>
 
 
